% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okf_graph.R
\name{okf_backlinks}
\alias{okf_backlinks}
\title{Concepts that link to a given concept ("linked from" / backlinks).}
\usage{
okf_backlinks(con, path)
}
\arguments{
\item{con}{An open DuckDB connection to an okf catalog.}

\item{path}{Bundle-relative concept path.}
}
\value{
Character vector of source concept paths (resolved inbound links).
}
\description{
Concepts that link to a given concept ("linked from" / backlinks).
}
